decompose(turn_bridge): S10 extract delivery epilogue (WIP, root 1071)
